+++ src/main/java/DAO/GameDAO.java
@@ -1,0 +1,21 @@
+package DAO;
+
+/**
+ * DAO class to store game instances.
+ * Whenever a game is called or started, store here. Once a game ends, remove it from table.
+ *
+ * @author Nathan (goodguyplayer)
+ * @Version 0
+ * @since 2021-04-18
+ */
+
+/*
+Changelog.:
+
+Version 0.:
+    - class created
+    - implemented
+ */
+
+public class GameDAO{
+}